# openapi.yaml
openapi: 3.1.0
info:
  title: komondor-api
  version: "1.0.0"
  summary: TSL Sequence Data API — projects, samples, sequencing runs and their files.
  description: |
    A description of what komondor-api **does today**, derived by reading
    `routes/` and `models/`. It is a record of current behaviour, not a
    proposal. Where the current behaviour is a known defect it is documented as
    it is, and the operation carries a `DRIFT` or `DEFECT` note saying so.
    Fixing one of those is a change to the API and must update this file in the
    same commit.

    `__tests__/contract/openapi.test.js` fails if a route exists in `routes/`
    with no entry here, so the two cannot silently diverge.

    ## Known drift from the sibling clients

    Three repositories consume this API: `komondor-web` (Nuxt 2 UI),
    `komondor-power` (bulk CSV ingest) and `komondor-nudge` (a scheduled
    worker that reads the same MongoDB directly rather than over HTTP). The
    following are the places where their assumptions and this server's
    behaviour have come apart. See `docs/CONTRACTS.md` for the policy.

    1. **Routes are mounted at the ROOT, not under `/api`.** `app.js` calls
       `app.use(authRoutes)` with no path prefix, so login is `POST /login`,
       not `POST /api/login` or `POST /auth/login`. The comment above the login
       handler in `routes/auth.js` still reads "Add POST - /api/login", which
       is where the belief came from. `komondor-power`'s client tries
...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /projects:
    get:
      tags: [Projects]
      summary: Projects visible to the caller.
      description: |
        Visibility comes from `buildVisibilityFilter`: full-access users and
        admins see everything, everyone else sees projects they own or that
        belong to one of their token's groups. A caller with neither a username
        nor groups matches nothing (BREAKING_CHANGES §1).

        Sorted by `createdAt` descending, in JavaScript rather than in Mongo.
        `group` is populated.
      responses:
        "200":
          description: Visible projects, newest first.
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "500":
          $ref: "#/components/responses/ServerError"
  /projects/names:
    get:
      tags: [Projects]
      summary: Every project name. Authenticated.
      description: |
        Requires a bearer token. This route previously had no
        `isAuthenticated` guard and handed every project name in the database
        to anonymous callers.

        The list is still UNFILTERED across groups, deliberately:
        `Project.name` carries a unique index, so a client needs the full list
        to tell a user that the name they have chosen is already taken. It
        returns names only — no ids, owners or group membership.
      responses:
        "200":
          description: All project names, across every group.
          content:
            application/json:
              schema:
                type: object
                properties:
                  projectNames:
                    type: array
                    items:
                      type: string
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "500":
          $ref: "#/components/responses/ServerError"
  /project:
    get:
      tags: [Projects]
      summary: One project, with samples, additional files and on-disk status.
      description: |
        The id is a **query parameter**, not a path segment.

        Alongside the project this returns a reconciliation of the database
        records against the `additional/` directory on disk. A datastore that
        cannot be read degrades to `additionalFilesStatus.status = "UNKNOWN"`
        rather than failing the request. Once a project has left `hpc`, the
        directory is not read: the listing is null and reconciliation is
        `NOT_APPLICABLE`.
      parameters:
        - name: id
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/ObjectId"
      responses:
        "200":
          description: The project and its file reconciliation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  project:
                    $ref: "#/components/schemas/Project"
                  location:
                    $ref: "#/components/schemas/EntityLocation"
                  actualAdditionalFiles:
                    type: [array, "null"]
                    items:
                      type: string
                    description: Filenames actually present in the project's `additional/` directory.
                  additionalFilesStatus:
                    $ref: "#/components/schemas/FileReconciliation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/GroupPermissionDenied"
        "404":
          $ref: "#/components/responses/NotFoundEnvelope"
        "500":
          $ref: "#/components/responses/ServerError"
  /project/toggle-nudgeable:
    put:
      tags: [Projects]
      summary: Set a project's nudgeable flag.
      description: |
        Requires WRITE access to the group that owns the project: the project
        is loaded first and the caller is authorised against its group, so a
        project outside the caller's writable groups answers 403.

        `nudgeable` must be a real JSON boolean — a non-boolean is a 400, not a
        truthy coercion, because `"false"` would otherwise set the opposite
        flag. This is stricter than `POST /projects/new`, which falls back to
        the group default rather than refusing.

        Answers `message` on success and does not return the updated document.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [_id, nudgeable]
              properties:
                _id:
                  $ref: "#/components/schemas/ObjectId"
                nudgeable:
                  type: boolean
      responses:
        "200":
          description: Updated. The project itself is not returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    const: Nudgeable status updated successfully.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/GroupPermissionDenied"
        "404":
          $ref: "#/components/responses/NotFoundEnvelope"
        "500":
          $ref: "#/components/responses/ServerError"
  /projects/new:
    post:
      tags: [Projects]
      summary: Create a project.
      description: |
        Requires WRITE access to the target group. A cross-group *reader*
        (`FULL_RECORDS_ACCESS_USERS`) cannot create a project in a group they
        are not a member of.

        `nudgeable` is HONOURED when the body carries a real JSON boolean, and
        otherwise DERIVED from the target group:

            nudgeable = (targetGroup.sendToEna === true)

        This replaced a hard-coded '2Blades' group ObjectId that was a stand-in
        for exactly this flag but only covered one of the two groups carrying
        it. A non-boolean `nudgeable` (for example the string `"true"`) is NOT
        a 400 here: it is discarded in favour of the group default, and the
        server logs `[projects/new] Ignoring non-boolean 'nudgeable'`.
        This API honours whatever real boolean the client sends. It does NOT
        currently mean `komondor-power`'s `project_nudgeable` CSV column
        reaches the database: `insertMetadata.ts` there builds the request as
        `nudgeable: !doNotSendToEna` and does not forward `project_nudgeable`
        under this field at all, so every project it creates still gets this
        API's group-derived default in practice, regardless of how that column
        serialises. Tracked as a Power-side gap in `docs/CONTRACTS.md` — this
        endpoint's own behaviour is not the drift, that consumer's request body
        is.

        Note that a group whose `sendToEna` is absent or `false` produces
        `nudgeable: false`. `PUT /project/toggle-nudgeable` sets the flag
        afterwards.

        NOT idempotent, unlike `/samples/new` and `/runs/new`: a repeated name
        hits the unique index on `Project.name` and fails. There is no 200
        branch here.

        On failure after the document is saved, the handler deletes it again —
        but files already moved are not cleaned up.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, group, shortDesc, longDesc, owner]
              properties:
                name:
                  type: string
                group:
                  $ref: "#/components/schemas/ObjectId"
                shortDesc:
                  type: string
                longDesc:
                  type: string
                owner:
                  type: string
                  description: A username, which need not correspond to an existing `User`.
                doNotSendToEna:
                  type: boolean
                doNotSendToEnaReason:
                  type: string
                nudgeable:
                  type: boolean
                  description: |
                    Honoured when sent as a real JSON boolean. Anything else
                    (including the strings `"true"` / `"false"`) is discarded
                    with a server-side warning and the group-derived default is
                    used instead. See the operation description.
                additionalFiles:
                  type: array
                  items:
                    type: object
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  project:
                    $ref: "#/components/schemas/Project"
        "400":
          description: Missing group, or mongoose validation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/GroupPermissionDenied"
        "500":
          $ref: "#/components/responses/ServerError"
  /samples:
    get:
      tags: [Samples]
      summary: Samples visible to the caller, newest first.
      description: Same visibility rules as `GET /projects`. Sorted in Mongo here.
      responses:
        "200":
          description: Visible samples.
          content:
            application/json:
              schema:
                type: object
                properties:
                  samples:
                    type: array
                    items:
                      $ref: "#/components/schemas/Sample"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        `Authorization: Bearer <token>`, signed with `JWT_SECRET` (HS256) by
        `lib/utils/jwtSign.js`.

        **The `exp` claim is mandatory.** `lib/utils/getUserFromRequest.js`
        rejects a structurally valid, correctly signed token that carries no
        `exp` by raising `TokenExpiredError`, which surfaces as 401. This was
        added deliberately (commit "refuse legacy tokens without an expiry
        claim"): tokens issued before August 2026 had no expiry and stayed
        valid forever, including tokens minted during the LDAP incident with an
        empty `groups` claim. Such tokens can only be fixed by re-logging in.

        Tokens expire after `JWT_EXPIRES_IN`, default `7d`.

        **Group membership is a snapshot.** `groups` is resolved once, at
        login, and never re-read. A user added to a group sees nothing new
        until their token expires and they sign in again — that expiry is the
        upper bound on the staleness.

        Requests with NO `Authorization` header are not rejected by the
        app-level middleware; they simply arrive with no `req.user`, and each
        route's `isAuthenticated` guard turns that into a 401. Routes without
        that guard (`/projects/names`, `GET /options/*`, `/health`, `/test`,
        `/login`, `/logout`, `/uploads`) serve anonymous callers.
...
  responses:
    Unauthenticated:
      description: |
        No usable token. Two distinct bodies reach clients as 401:
        `{ error: "Authentication required" }` from a route's
        `isAuthenticated` guard, and
        `{ error: "Invalid or expired authentication token", detail }` from the
        app-level middleware in `app.js` when a token was present but bad or
        expired.
      content:
        application/json:
          schema:
            oneOf:
              - type: object
                properties:
                  error:
                    const: Authentication required
              - type: object
                properties:
                  error:
                    const: Invalid or expired authentication token
                  detail:
                    type: string
                    description: The underlying jsonwebtoken message.
                  requestId:
                    type: string
                    description: Present only when raised by the terminal error handler.
...
    GroupPermissionDenied:
      description: |
        The caller is neither the record's owner, nor a member of its group,
        nor an admin. The message embeds the username:
        `User '<username>' does not have permission to ...`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    BadRequest:
      description: A required parameter is missing or malformed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    NotFoundEnvelope:
      description: The referenced record does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    ServerError:
      description: |
        An unhandled failure. In production a 500 replaces `error` with a
        generic string but STILL includes `detail` — komondor-power is an
        internal client and relies on it for diagnostics.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
    ObjectId:
      type: string
      pattern: "^[0-9a-fA-F]{24}$"
      # Deliberately not the old hard-coded '2Blades' group id: that literal
      # used to carry meaning in routes/projects.js and reusing it as a generic
      # example invited the reader to think it still does.
      examples: ["507f1f77bcf86cd799439011"]
    ErrorEnvelope:
      type: object
      description: |
        The standard failure body, produced by `handleError` in
        `routes/_utils.js` and by the terminal handler in `app.js`.

        `detail` carries the underlying error message — more specific than
        `error` (e.g. "E11000 duplicate key error" against "Failed to create
        new project.") — and is present even for production 500s.

        Not universal: `/login` uses `message`; `/groups/*` and `/me` send a
        bare `error`; `/read-file` and `/directory-files` send `{ error }` with
        a 200.
      properties:
        error:
          type: string
          description: A user-facing message.
        detail:
          type: string
          description: The underlying error message, for diagnostics.
        requestId:
          type: string
          description: Correlates with the server log line for this failure.
          examples: ["1735689600000-k3j9x2p1a"]
        code:
          type: string
          description: Optional stable machine-readable error code.
        projectId:
          type: string
          description: Present on project-storage refusals.
        storageState:
          type: string
          description: Present on project-storage refusals; may be `invalid`.
...
    EntityLocation:
      type: object
      description: Display-ready authoritative location for a detail response.
      properties:
        authoritative:
          type: [string, "null"]
          enum: [hpc, s3, null]
        baseUri:
          type: [string, "null"]
        rawUri:
          type: [string, "null"]
          description: Present on run details only.
        additionalUri:
          type: [string, "null"]
        plannedS3Uri:
          type: [string, "null"]
          description: Set while migrating; it is not yet authoritative.
...
    Group:
      type: object
      properties:
        _id:
          $ref: "#/components/schemas/ObjectId"
        name:
          type: string
        safeName:
          type: string
          description: Filesystem-safe name; also the group's directory under `DATASTORE_ROOT`.
        ldapGroups:
          type: array
          description: LDAP `memberOf` strings mapping onto this group. Matched case-insensitively.
          items:
            type: string
        deleted:
          type: boolean
          default: false
          description: |
            SOFT-DELETE FLAG. `POST /groups/delete` sets it; `POST
            /groups/resurrect` clears it. The document, its directory and every
            record referencing it survive.

            Consumers have had to filter on this themselves — `komondor-nudge`
            declares the field on its read-only model and records
            `GROUP_DELETED` when it skips a project for this reason.
        sendToEna:
          type: boolean
          default: false
        oldId:
          type: string
          description: Migration artefact.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Project:
      type: object
      properties:
        _id:
          $ref: "#/components/schemas/ObjectId"
        name:
          type: string
          description: Unique across the collection.
        safeName:
          type: string
        owner:
          type: string
        shortDesc:
          type: string
        longDesc:
          type: string
        group:
          description: An ObjectId, or the populated `Group` on endpoints that populate it.
          oneOf:
            - $ref: "#/components/schemas/ObjectId"
            - $ref: "#/components/schemas/Group"
        path:
          type: string
          description: Datastore-relative path, `/<group.safeName>/<project.safeName>`.
        isPublic:
          type: boolean
          default: false
        accessions:
          type: array
          items:
            type: string
        releaseDate:
          type: string
          description: "ENA submission date, stored as a `DD-MM-YYYY` STRING, not a date."
        doNotSendToEna:
          type: boolean
          default: false
        doNotSendToEnaReason:
          type: string
        nudgeable:
          type: boolean
          default: true
          description: |
            GROUP-DERIVED DEFAULT ON CREATE. `POST /projects/new` honours a
            real JSON boolean from the client and otherwise computes
            `nudgeable = (targetGroup.sendToEna === true)`. A group with
            `sendToEna` absent or `false` therefore produces `nudgeable:
            false`. `PUT /project/toggle-nudgeable` sets the flag afterwards,
            and requires write access to the project's group.

            `komondor-nudge` reads this field directly from MongoDB to decide
            whether to email a project owner. A project where the field is
            ABSENT (predating the field) is skipped and recorded as
            `NUDGEABLE_UNDEFINED` unless
            `NUDGE_INCLUDE_UNDEFINED_NUDGEABLE=true` — so absent and `false`
            are not the same thing to that consumer, even though the API can no
            longer produce an absent value.
        storage:
          $ref: "#/components/schemas/StorageSummary"
        nudges:
          type: array
          description: |
            Timestamps of nudge emails sent. Written by komondor-nudge directly
            against MongoDB (`$push`), never through this API.
          items:
            type: string
            format: date-time
        samples:
          type: array
          description: Virtual. Populated by `GET /project` only.
          items:
            $ref: "#/components/schemas/Sample"
        additionalFiles:
          type: array
          description: Virtual. Populated by `GET /project` only.
          items:
            type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    FileReconciliation:
      type: object
      description: |
        The result of comparing a record's file documents against the files
        actually on disk, returned by `GET /project`, `GET /sample` and
        `GET /run`.

        Comparison is by COUNT per filename, not by set membership, so two
        records named `report.pdf` with one copy on disk is reported as
        missing one. Names are compared after NFC normalisation, because macOS
        and Linux disagree on how accented filenames are encoded.
      properties:
        status:
          type: string
          enum: [OK, WARNING, MISMATCH, UNKNOWN, NOT_APPLICABLE]
          description: |
            `MISMATCH` — files are missing from disk, or a record's filename
            could not be resolved. `WARNING` — only untracked extra files on
            disk. `UNKNOWN` — the directory could not be read at all; the
            request still succeeds rather than 500-ing on an unmounted
            datastore. `NOT_APPLICABLE` means the project has left live HPC
            storage, so no filesystem check was attempted.
        reason:
          type: string
          description: "PROJECT_STORAGE_READ_ONLY for NOT_APPLICABLE."
        storageState:
          type: string
          description: Project storage state for `NOT_APPLICABLE`.
        message:
          type: string
          description: A human-readable summary. A file both missing and extra is reported as a probable rename.
        missing:
          type: array
          description: Named by a record but absent from disk. Repeated by the shortfall.
          items:
            type: string
        extra:
          type: array
          description: Present on disk but not in any record. Hidden files and in-flight `.part-` copies are excluded.
          items:
            type: string
        unresolved:
          type: array
          description: |
            Ids of records whose filename could not be determined — usually an
            unpopulated ref, or a `File` deleted out from under its
            `AdditionalFile`.
          items:
            type: string